requirements.txt
huggingface_hub
python-dotenv
Pillow
diffusers
torch
torchvision
accelerate
transformers
python-telegram-bot
asyncio